Consolidate page UI compoent files
